Attention world builders:
This commit stabilizes zoning through traditional zonelines a bit, but it looks like our data is off. zone_line_nodes in the included SQL is where the server gets its coords from and needs some tweaking. It also includes PEQ's zone and zone_points tables with the unsupported zones removed. zone_points can probably be used to get proper zone-in, and in some cases zone-out coords if needed. This SQL is REQUIRED.

Also added some logging and testing code for doors to try to get them to work :I