Corrected the problem preventing zones with lower zoneids from booting up. Stabilized zoning a bit (Every now and again a zone will fail to bootup - this is a problem inherent to older EQEmu versions. It'll be corrected once the EQEmu 0.8 zoning code is ported over.) #zone now accepts coords, and zoneidnumber. Renamed #goto to #moveto hopefully to get around the duplicate command bug until the cause can be found.
To correct the Surefall to Qeynos Hills zoneline, issue this SQL:
update zone_line_nodes set target_x = 196.7, target_y = 5100.9, target_z = -1 where zone = 'qrg' and target_zone = 'qeytoqrg';

Many zones will need to be repaired like that, as the in-coords are off by a few and cause you to go off the map. The EQEmu 0.8 zone_points are correct and can be used.